extracting time tracking into hook
