+++ src/utils/constants.ts
@@ -1,6 +1,6 @@
 export const CHAT_UI_REOPEN = "___chatmate-chat-widget-open___";
-export const API_PATH = "https://chatmate.so";
-// export const API_PATH = "http://localhost:3000";
+// export const API_PATH = "https://chatmate.so";
+export const API_PATH = "http://localhost:3000";
 
 export const colors = {
   zinc: "zinc-900",
